# backend/openapi/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Liveprofit API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/analysis-tasks/{task_id}/quant-signals": {
      "get": {
        "tags": [
          "quant-signals"
        ],
        "summary": "List Quant Signals",
        "operationId": "list_quant_signals_api_v1_analysis_tasks__task_id__quant_signals_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "buy",
                "holding",
                "orders",
                "errors"
              ],
              "type": "string",
              "default": "buy",
              "title": "Kind"
            }
          },
          {
            "name": "attempt_no",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Attempt No"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_QuantSignalPageData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EnvelopeMeta": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "请求标识（X-Trace-ID）"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "v1"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "不透明续页游标；无下一页为 null"
          }
        },
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "EnvelopeMeta"
      },
...
      "Envelope_QuantSignalPageData_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/QuantSignalPageData"
          },
          "meta": {
            "$ref": "#/components/schemas/EnvelopeMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[QuantSignalPageData]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuantSignalPageData": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QuantSignalRowDTO"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "attempt_no": {
            "type": "integer",
            "title": "Attempt No"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          }
        },
        "type": "object",
        "required": [
          "items",
          "attempt_no",
          "kind"
        ],
        "title": "QuantSignalPageData"
      },
      "QuantSignalRowDTO": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "ts_code": {
            "type": "string",
            "title": "Ts Code"
          },
          "signal_kind": {
            "type": "string",
            "title": "Signal Kind"
          },
          "attempt_no": {
            "type": "integer",
            "title": "Attempt No"
          },
          "action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "entry_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Price"
          },
          "stop_loss": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Loss"
          },
          "take_profit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Take Profit"
          },
          "sell_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sell Ratio"
          },
          "order_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Status"
          },
          "shares": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shares"
          },
          "notional": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notional"
          },
          "order_cost_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Cost Price"
          },
          "valuation_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valuation Price"
          },
          "risk_bucket": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Bucket"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ts_code",
          "signal_kind",
          "attempt_no"
        ],
        "title": "QuantSignalRowDTO"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}